fixed the transaction problem and generated the Jasper report

+++ src/main/java/lk/ijse/oxford/contoller/student_form_controllers/DeleteStudentFormController.java
@@ -1,19 +1,30 @@
 package lk.ijse.oxford.contoller.student_form_controllers;
 
 import com.jfoenix.controls.JFXButton;
+import com.jfoenix.controls.JFXTextField;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.Cursor;
+import javafx.scene.Node;
+import javafx.scene.chart.BarChart;
+import javafx.scene.chart.XYChart;
 import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
+import javafx.scene.input.KeyEvent;
 import javafx.scene.input.MouseEvent;
+import lk.ijse.oxford.db.DbConnection;
 import lk.ijse.oxford.model.Student;
 import lk.ijse.oxford.model.tm.AttedanceTm;
 import lk.ijse.oxford.model.tm.StudentTm;
 import lk.ijse.oxford.repository.StudentRepo;
+import lk.ijse.oxford.util.Regex;
+import lk.ijse.oxford.util.TextFields;
 
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
@@ -21,12 +32,14 @@ import java.util.Optional;
 
 public class DeleteStudentFormController {
     @FXML
+    private BarChart bcStudentChart;
+    @FXML
     private Label lblStudentCount;
     private int studentCount;
     @FXML
-    private TextField txtStudentId;
+    private JFXTextField txtStudentId;
     @FXML
-    private TextField txtStudentName;
+    private JFXTextField txtStudentName;
     @FXML
     private TableColumn<?,?> colStId;
     @FXML
@@ -47,6 +60,11 @@ public class DeleteStudentFormController {
         this.studentList = getAllStudents();
         setCellValueFactory();
         loadStudentTable();
+        try {
+            studentAttendance(bcStudentChart);
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
         try {
             studentCount = StudentRepo.getStudentCount();
         } catch (SQLException e) {
@@ -96,27 +114,60 @@ public class DeleteStudentFormController {
         colUserId.setCellValueFactory(new PropertyValueFactory<>("userId"));
     }
     public void btnStudentDeleteOnAction(ActionEvent actionEvent) {
-        String id = txtStudentId.getText();
+        if (isValidate()){
+            String id = txtStudentId.getText();
 
-        try {
-            boolean isDeleted = StudentRepo.delete(id);
-            if (isDeleted) {
-                new Alert(Alert.AlertType.CONFIRMATION, "Student Data Deleted!").show();
+            try {
+                boolean isDeleted = StudentRepo.delete(id);
+                if (isDeleted) {
+                    new Alert(Alert.AlertType.CONFIRMATION, "Student Data Deleted!").show();
+                    initialize();
+                }
+            } catch (SQLException e) {
+                new Alert(Alert.AlertType.ERROR, e.getMessage()).show();
             }
-        } catch (SQLException e) {
-            new Alert(Alert.AlertType.ERROR, e.getMessage()).show();
         }
-    }
-
-    public void btnStudentRefreshOnAction(ActionEvent actionEvent) {
-        initialize();
-        txtStudentName.setText("");
-        txtStudentId.setText("");
     }
 
     public void deleteClickedOnAction(MouseEvent mouseEvent) {
         StudentTm selectedItem = tblStudent.getSelectionModel().getSelectedItem();
         txtStudentId.setText(selectedItem.getStId());
         txtStudentName.setText(selectedItem.getName());
+    }
+
+    public boolean isValidate(){
+        if(!Regex.setTextColor(TextFields.NAME,txtStudentName))return false;
+        if(!Regex.setTextColor(TextFields.SID,txtStudentId))return false;
+        return true;
+    }
+    public void txtStIdCheckOnAction(KeyEvent keyEvent) {
+        Regex.setTextColor(TextFields.SID,txtStudentId);
+    }
+    public void txtStNameCheckOnAction(KeyEvent keyEvent) {
+        Regex.setTextColor(TextFields.NAME,txtStudentName);
+    }
+
+    private void studentAttendance(BarChart<String , Number> bcStudentChart) throws SQLException {
+        Connection connection = DbConnection.getInstance().getConnection();
+
+        String sql = "select MONTHNAME(date) as month, count(AttendMark) attendMark from Attendance group by MONTHNAME(date);";
+
+        PreparedStatement pstm = connection.prepareStatement(sql);
+
+        ResultSet resultSet = pstm.executeQuery();
+
+        XYChart.Series<String, Number> series = new XYChart.Series<>();
+
+        while (resultSet.next()) {
+            String month = resultSet.getString("month");
+            int attendMark = resultSet.getInt("attendMark");
+            series.getData().add(new XYChart.Data<>(month, attendMark));
+        }
+
+        bcStudentChart.getData().add(series);
+
+        for(Node n:bcStudentChart.lookupAll(".default-color0.chart-bar")) {
+            n.setStyle("-fx-bar-fill: #0C87F2;");
+        }
     }
 }
